CSWConstraint

CSWConstraint

CSWConstraint

Format

R6Class object.

Value

Object of R6Class for modelling an CSW Constraint

Author

Emmanuel Blondel <emmanuel.blondel1@gmail.com>

Super class

ows4R::OGCAbstractObject -> CSWConstraint

Public fields

wrap

internal property for object XML encoding

CqlText

text to use as CQL filter

filter

Methods

Public methods

CSWConstraint$new()

CSWConstraint$setServiceVersion()

CSWConstraint$clone()

Inherited methods

ows4R::OGCAbstractObject$ERROR()

ows4R::OGCAbstractObject$INFO()

ows4R::OGCAbstractObject$WARN()

ows4R::OGCAbstractObject$encode()

ows4R::OGCAbstractObject$getClass()

ows4R::OGCAbstractObject$getClassName()

ows4R::OGCAbstractObject$getNamespaceDefinition()

ows4R::OGCAbstractObject$isFieldInheritedFrom()

ows4R::OGCAbstractObject$logger()

ows4R::OGCAbstractObject$print()

Method new()

Initializes a CSWConstraint object to be used to constrain CSW operations.

Usage

CSWConstraint$new(cqlText = NULL, filter = NULL, serviceVersion = "2.0.2")

Arguments

cqlText

cqlText, object of class character

filter

filter, object extending OGCFilter

serviceVersion

CSW service version. Default is "2.0.2"

Method setServiceVersion()

Set service version. This methods ensures that underlying filter property is properly set with the right OGC filter version.

Usage

CSWConstraint$setServiceVersion(serviceVersion)

Arguments

serviceVersion

service version

Method clone()

The objects of this class are cloneable with this method.

Usage

CSWConstraint$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

  filter <- OGCFilter$new( PropertyIsEqualTo$new("apiso:Identifier", "12345") )
  cons <- CSWConstraint$new(filter = filter)
  cons_xml <- cons$encode() #how it looks like in XML