WCSCoverageSummary

WCSCoverageSummary

WCSCoverageSummary

Format

R6Class object.

Value

Object of R6Class modelling a WCS coverage summary

Note

Class used internally by ows4R.

Author

Emmanuel Blondel <emmanuel.blondel1@gmail.com>

Super class

ows4R::OGCAbstractObject -> WCSCoverageSummary

Public fields

CoverageId

coverage id

CoverageSubtype

coverage subtype

CoverageSubtypeParent

coverage subtype parent

WGS84BoundingBox

WGS84 bounding box

BoundingBox

bounding box

Methods

Public methods

WCSCoverageSummary$new()

WCSCoverageSummary$getId()

WCSCoverageSummary$getSubtype()

WCSCoverageSummary$getSubtypeParent()

WCSCoverageSummary$getWGS84BoundingBox()

WCSCoverageSummary$getBoundingBox()

WCSCoverageSummary$getDescription()

WCSCoverageSummary$getDimensions()

WCSCoverageSummary$getCoverage()

WCSCoverageSummary$getCoverageStack()

WCSCoverageSummary$clone()

Inherited methods

ows4R::OGCAbstractObject$ERROR()

ows4R::OGCAbstractObject$INFO()

ows4R::OGCAbstractObject$WARN()

ows4R::OGCAbstractObject$encode()

ows4R::OGCAbstractObject$getClass()

ows4R::OGCAbstractObject$getClassName()

ows4R::OGCAbstractObject$getNamespaceDefinition()

ows4R::OGCAbstractObject$isFieldInheritedFrom()

ows4R::OGCAbstractObject$logger()

Method new()

Initializes a WCSCoverageSummary object

Usage

WCSCoverageSummary$new(
  xmlObj,
  capabilities,
  serviceVersion,
  owsVersion,
  logger = NULL
)

Arguments

xmlObj

object of class XMLInternalNode-class from XML

capabilities

object of class WCSCapabilities

serviceVersion

WCS service version

owsVersion

version

logger

logger type NULL, "INFO" or "DEBUG"

Method getId()

Get coverage ID

Usage

WCSCoverageSummary$getId()

Returns

an object of class character

Method getSubtype()

Get sub type

Usage

WCSCoverageSummary$getSubtype()

Returns

an object of class character

Method getSubtypeParent()

Get sub type parent

Usage

WCSCoverageSummary$getSubtypeParent()

Returns

an object of class character

Method getWGS84BoundingBox()

Get bounding box

Usage

WCSCoverageSummary$getWGS84BoundingBox()

Returns

an object of class OWSWGS84BoundingBox

Method getBoundingBox()

Get WGS84 bounding box

Usage

WCSCoverageSummary$getBoundingBox()

Returns

an object of class OWSBoundingBox

Method getDescription()

Get description

Usage

WCSCoverageSummary$getDescription()

Returns

an object of class WCSCoverageDescription

Method getDimensions()

Get dimensions

Usage

WCSCoverageSummary$getDimensions()

Returns

the list of dimensions

Method getCoverage()

Get coverage data

Usage

WCSCoverageSummary$getCoverage(
  bbox = NULL,
  crs = NULL,
  time = NULL,
  elevation = NULL,
  format = NULL,
  rangesubset = NULL,
  gridbaseCRS = NULL,
  gridtype = NULL,
  gridCS = NULL,
  gridorigin = NULL,
  gridoffsets = NULL,
  method = "GET",
  filename = NULL,
  ...
)

Arguments

bbox

bbox. Default is NULL

crs

crs. Default is NULL

time

time. Default is NULL

elevation

elevation. Default is NULL

format

format. Default will be GeoTIFF, coded differently depending on the WCS version.

rangesubset

rangesubset. Default is NULL

gridbaseCRS

grid base CRS. Default is NULL

gridtype

grid type. Default is NULL

gridCS

grid CS. Default is NULL

gridorigin

grid origin. Default is NULL

gridoffsets

grid offsets. Default is NULL

method

method to get coverage, either 'GET' or 'POST' (experimental - under development)

filename

filename. Optional filename to download the coverage

...

any other argument to WCSGetCoverage

Returns

an object of class SpatRaster from terra

Method getCoverageStack()

Get a spatio-temporal coverage data cubes as coverage stack

Usage

WCSCoverageSummary$getCoverageStack(
  time = NULL,
  elevation = NULL,
  bbox = NULL,
  filename_handler = NULL,
  ...
)

Arguments

time

time

elevation

elevation

bbox

bbox

filename_handler

Optional filename handling function with arguments 'identifier', 'time', 'elevation', 'bbox', 'format' See WCSCoverageFilenameHandler as genric filename handler that can be used.

...

any other parameter to pass to getCoverage

Returns

an object of class stack from raster

Method clone()

The objects of this class are cloneable with this method.

Usage

WCSCoverageSummary$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.