WFSCapabilities

WFSCapabilities

WFSCapabilities

Format

R6Class object.

Value

Object of R6Class with methods for interfacing an OGC Web Feature Service Get Capabilities document.

Note

Class used to read a WFSCapabilities document. The use of WFSClient is recommended instead to benefit from the full set of capabilities associated to a WFS server.

Author

Emmanuel Blondel <emmanuel.blondel1@gmail.com>

Super classes

ows4R::OGCAbstractObject -> ows4R::OWSCapabilities -> WFSCapabilities

Methods

Public methods

WFSCapabilities$new()

WFSCapabilities$getFeatureTypes()

WFSCapabilities$findFeatureTypeByName()

WFSCapabilities$clone()

Inherited methods

ows4R::OGCAbstractObject$ERROR()

ows4R::OGCAbstractObject$INFO()

ows4R::OGCAbstractObject$WARN()

ows4R::OGCAbstractObject$encode()

ows4R::OGCAbstractObject$getClass()

ows4R::OGCAbstractObject$getClassName()

ows4R::OGCAbstractObject$getNamespaceDefinition()

ows4R::OGCAbstractObject$isFieldInheritedFrom()

ows4R::OGCAbstractObject$logger()

ows4R::OGCAbstractObject$print()

ows4R::OWSCapabilities$getClient()

ows4R::OWSCapabilities$getOWSVersion()

ows4R::OWSCapabilities$getOperationsMetadata()

ows4R::OWSCapabilities$getRequest()

ows4R::OWSCapabilities$getService()

ows4R::OWSCapabilities$getServiceIdentification()

ows4R::OWSCapabilities$getServiceProvider()

ows4R::OWSCapabilities$getServiceVersion()

ows4R::OWSCapabilities$getUrl()

ows4R::OWSCapabilities$setClient()

Method new()

Initializes a WFSCapabilities object

Usage

WFSCapabilities$new(url, version, logger = NULL, ...)

Arguments

url

url

version

version

logger

logger type NULL, "INFO" or "DEBUG"

...

any other parameter to pass to OWSGetCapabilities service request

Method getFeatureTypes()

List the feature types available. If pretty is TRUE, the output will be an object of class data.frame

Usage

WFSCapabilities$getFeatureTypes(pretty = FALSE)

Arguments

pretty

whether the output should be summarized as data.frame

Returns

a list of WFSFeatureType or a data.frame

Method findFeatureTypeByName()

Finds a feature type by name

Usage

WFSCapabilities$findFeatureTypeByName(expr, exact = TRUE)

Arguments

expr

expr

exact

exact matching? Default is TRUE

Method clone()

The objects of this class are cloneable with this method.

Usage

WFSCapabilities$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.