WFSFeatureType

WFSFeatureType

WFSFeatureType

Format

R6Class object.

Value

Object of R6Class modelling a WFS feature type

Note

Class used internally by ows4R to trigger a WFS DescribeFeatureType request

Author

Emmanuel Blondel <emmanuel.blondel1@gmail.com>

Super class

ows4R::OGCAbstractObject -> WFSFeatureType

Public fields

description

description

features

features

Methods

Public methods

WFSFeatureType$new()

WFSFeatureType$getName()

WFSFeatureType$getTitle()

WFSFeatureType$getAbstract()

WFSFeatureType$getKeywords()

WFSFeatureType$getDefaultCRS()

WFSFeatureType$getBoundingBox()

WFSFeatureType$getDescription()

WFSFeatureType$hasGeometry()

WFSFeatureType$getGeometryType()

WFSFeatureType$getFeaturesCRS()

WFSFeatureType$getFeatures()

WFSFeatureType$clone()

Inherited methods

ows4R::OGCAbstractObject$ERROR()

ows4R::OGCAbstractObject$INFO()

ows4R::OGCAbstractObject$WARN()

ows4R::OGCAbstractObject$encode()

ows4R::OGCAbstractObject$getClass()

ows4R::OGCAbstractObject$getClassName()

ows4R::OGCAbstractObject$getNamespaceDefinition()

ows4R::OGCAbstractObject$isFieldInheritedFrom()

ows4R::OGCAbstractObject$logger()

Method new()

Initializes an object of class WFSFeatureType

Usage

WFSFeatureType$new(xmlObj, capabilities, version, logger = NULL)

Arguments

xmlObj

an object of class XMLInternalNode-class to initialize from XML

capabilities

object of class WFSCapabilities

version

service version

logger

logger

Method getName()

Get feature type name

Usage

WFSFeatureType$getName()

Arguments

object

of class character

Method getTitle()

Get feature type title

Usage

WFSFeatureType$getTitle()

Arguments

object

of class character

Method getAbstract()

Get feature type abstract

Usage

WFSFeatureType$getAbstract()

Arguments

object

of class character

Method getKeywords()

Get feature type keywords

Usage

WFSFeatureType$getKeywords()

Arguments

object

of class character

Method getDefaultCRS()

Get feature type default CRS

Usage

WFSFeatureType$getDefaultCRS()

Arguments

object

of class character

Method getBoundingBox()

Get feature type bounding box

Usage

WFSFeatureType$getBoundingBox()

Arguments

object

of class matrix

Method getDescription()

Describes a feature type

Usage

WFSFeatureType$getDescription(pretty = FALSE)

Arguments

pretty

pretty whether to return a prettified data.frame. Default is FALSE

Returns

a list of WFSFeatureTypeElement or data.frame

Method hasGeometry()

Indicates with feature type has a geometry

Usage

WFSFeatureType$hasGeometry()

Returns

object of class logical

Method getGeometryType()

Get geometry type

Usage

WFSFeatureType$getGeometryType()

Returns

object of class character representing the geometry tpe

Method getFeaturesCRS()

Inherits features CRS

Usage

WFSFeatureType$getFeaturesCRS(obj)

Arguments

obj

features object

Returns

object of class integer

Method getFeatures()

Get features

Usage

WFSFeatureType$getFeatures(
  ...,
  outputFormat = NULL,
  paging = FALSE,
  paging_length = 1000,
  parallel = FALSE,
  parallel_handler = NULL,
  cl = NULL
)

Arguments

...

any other parameter to pass to the WFSGetFeature request

outputFormat

output format

paging

paging. Default is FALSE

paging_length

number of features to request per page. Default is 1000

parallel

whether to get features using parallel multicore strategy. Default is FALSE

parallel_handler

Handler function to parallelize the code. eg mclapply

cl

optional cluster object for parallel cluster approaches using eg. parallel::makeCluster

typeName

the name of the feature type

Returns

features as object of class sf

Method clone()

The objects of this class are cloneable with this method.

Usage

WFSFeatureType$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.