WMSCapabilities

WMSCapabilities

WMSCapabilities

Format

R6Class object.

Value

Object of R6Class with methods for interfacing an OGC Web Map Service Get Capabilities document.

Note

Class used to read a WMSCapabilities document. The use of WMSClient is recommended instead to benefit from the full set of capabilities associated to a WMS server.

Author

Emmanuel Blondel <emmanuel.blondel1@gmail.com>

Super classes

ows4R::OGCAbstractObject -> ows4R::OWSCapabilities -> WMSCapabilities

Methods

Public methods

WMSCapabilities$new()

WMSCapabilities$getRequests()

WMSCapabilities$getRequestNames()

WMSCapabilities$getLayers()

WMSCapabilities$findLayerByName()

WMSCapabilities$clone()

Inherited methods

ows4R::OGCAbstractObject$ERROR()

ows4R::OGCAbstractObject$INFO()

ows4R::OGCAbstractObject$WARN()

ows4R::OGCAbstractObject$encode()

ows4R::OGCAbstractObject$getClass()

ows4R::OGCAbstractObject$getClassName()

ows4R::OGCAbstractObject$getNamespaceDefinition()

ows4R::OGCAbstractObject$isFieldInheritedFrom()

ows4R::OGCAbstractObject$logger()

ows4R::OWSCapabilities$getClient()

ows4R::OWSCapabilities$getOWSVersion()

ows4R::OWSCapabilities$getOperationsMetadata()

ows4R::OWSCapabilities$getRequest()

ows4R::OWSCapabilities$getService()

ows4R::OWSCapabilities$getServiceIdentification()

ows4R::OWSCapabilities$getServiceProvider()

ows4R::OWSCapabilities$getServiceVersion()

ows4R::OWSCapabilities$getUrl()

ows4R::OWSCapabilities$setClient()

Method new()

Initializes a WMSCapabilities object

Usage

WMSCapabilities$new(url, version, logger = NULL, ...)

Arguments

url

url

version

version

logger

logger type NULL, "INFO" or "DEBUG"

...

any other parameter to pass to OWSGetCapabilities service request

Method getRequests()

List the requests available. If pretty is TRUE, the output will be an object of class data.frame

Usage

WMSCapabilities$getRequests(pretty = FALSE)

Arguments

pretty

pretty

Returns

a list of OWSRequest available, or a data.frame

Method getRequestNames()

List the names of requests available.

Usage

WMSCapabilities$getRequestNames()

Returns

object of class character

Method getLayers()

List the layers available. If pretty is TRUE, the output will be an object of class data.frame

Usage

WMSCapabilities$getLayers(pretty = FALSE)

Arguments

pretty

pretty

Returns

a list of WMSLayer available, or a data.frame

Method findLayerByName()

Finds a layer by name

Usage

WMSCapabilities$findLayerByName(expr, exact = TRUE)

Arguments

expr

expr

exact

exact matching? Default is TRUE

Method clone()

The objects of this class are cloneable with this method.

Usage

WMSCapabilities$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.