WPSCapabilities

WPSCapabilities

WPSCapabilities

Format

R6Class object.

Value

Object of R6Class with methods for interfacing an OGC Web Processing Service (WPS) Get Capabilities document.

Note

Class used to read a WPSCapabilities document. The use of WPSClient is recommended instead to benefit from the full set of capabilities associated to a WPS server.

Author

Emmanuel Blondel <emmanuel.blondel1@gmail.com>

Super classes

ows4R::OGCAbstractObject -> ows4R::OWSCapabilities -> WPSCapabilities

Methods

Public methods

WPSCapabilities$new()

WPSCapabilities$getProcesses()

WPSCapabilities$describeProcess()

WPSCapabilities$execute()

WPSCapabilities$clone()

Inherited methods

ows4R::OGCAbstractObject$ERROR()

ows4R::OGCAbstractObject$INFO()

ows4R::OGCAbstractObject$WARN()

ows4R::OGCAbstractObject$encode()

ows4R::OGCAbstractObject$getClass()

ows4R::OGCAbstractObject$getClassName()

ows4R::OGCAbstractObject$getNamespaceDefinition()

ows4R::OGCAbstractObject$isFieldInheritedFrom()

ows4R::OGCAbstractObject$logger()

ows4R::OWSCapabilities$getClient()

ows4R::OWSCapabilities$getOWSVersion()

ows4R::OWSCapabilities$getOperationsMetadata()

ows4R::OWSCapabilities$getRequest()

ows4R::OWSCapabilities$getService()

ows4R::OWSCapabilities$getServiceIdentification()

ows4R::OWSCapabilities$getServiceProvider()

ows4R::OWSCapabilities$getServiceVersion()

ows4R::OWSCapabilities$getUrl()

ows4R::OWSCapabilities$setClient()

Method new()

Initializes a WPSCapabilities object

Usage

WPSCapabilities$new(url, version, client = NULL, logger = NULL, ...)

Arguments

url

url

version

version

client

an object of class WPSClient

logger

logger type NULL, "INFO" or "DEBUG"

...

any other parameter to pass to OWSGetCapabilities service request

Method getProcesses()

Get the list of processes offered by the service capabilities. pretty allows to control the type output. If TRUE, a data.frame will be returned. When prettified output, it is also possible to get a 'full' description of the process by setting full = TRUE in which case a the WPS client will request a process description (with more information about the process) for each process listed in the capabilities.

Usage

WPSCapabilities$getProcesses(pretty = FALSE, full = FALSE)

Arguments

pretty

pretty

full

full

Returns

a list of WPSProcessDescription or a data.frame

Method describeProcess()

Get the description of a process, given its identifier, returning an object of class WPSProcessDescription

Usage

WPSCapabilities$describeProcess(identifier)

Arguments

identifier

process identifier

Returns

an object of class WPSProcessDescription

Method execute()

Execute a process, given its identifier

Usage

WPSCapabilities$execute(
  identifier,
  dataInputs = list(),
  responseForm = NULL,
  storeExecuteResponse = FALSE,
  lineage = NULL,
  status = NULL,
  update = FALSE,
  updateInterval = 1
)

Arguments

identifier

process identifier

dataInputs

a named list of data inputs, objects of class WPSLiteralData, WPSComplexData or WPSBoundingBoxData

responseForm

response form, object of class WPSResponseDocument

storeExecuteResponse

store execute response? object of class logical. FALSE by default

lineage

lineage, object of class logical

status

status, object of class logical

update

update, object of class logical. For asynchronous requests

updateInterval

update interval, object of class integer. For asynchronous requests

Method clone()

The objects of this class are cloneable with this method.

Usage

WPSCapabilities$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.