WPSClient

WPSClient

WPSClient

Format

R6Class object.

Value

Object of R6Class with methods for interfacing an OGC Web Processing Service.

Author

Emmanuel Blondel <emmanuel.blondel1@gmail.com>

Super classes

ows4R::OGCAbstractObject -> ows4R::OWSClient -> WPSClient

Methods

Public methods

WPSClient$new()

WPSClient$getCapabilities()

WPSClient$reloadCapabilities()

WPSClient$getProcesses()

WPSClient$describeProcess()

WPSClient$execute()

WPSClient$clone()

Inherited methods

ows4R::OGCAbstractObject$ERROR()

ows4R::OGCAbstractObject$INFO()

ows4R::OGCAbstractObject$WARN()

ows4R::OGCAbstractObject$encode()

ows4R::OGCAbstractObject$getClass()

ows4R::OGCAbstractObject$getClassName()

ows4R::OGCAbstractObject$getNamespaceDefinition()

ows4R::OGCAbstractObject$isFieldInheritedFrom()

ows4R::OGCAbstractObject$logger()

ows4R::OGCAbstractObject$print()

ows4R::OWSClient$getCASUrl()

ows4R::OWSClient$getConfig()

ows4R::OWSClient$getHeaders()

ows4R::OWSClient$getPwd()

ows4R::OWSClient$getToken()

ows4R::OWSClient$getUrl()

ows4R::OWSClient$getUser()

ows4R::OWSClient$getVersion()

Method new()

This method is used to instantiate a WPSClient with the url of the OGC service. Authentication is supported using basic auth (using user/pwd arguments), bearer token (using token argument), or custom (using headers argument). By default, the logger argument will be set to NULL (no logger). This argument accepts two possible values: INFO: to print only ows4R logs, DEBUG: to print more verbose logs

Usage

WPSClient$new(
  url,
  serviceVersion = NULL,
  user = NULL,
  pwd = NULL,
  token = NULL,
  headers = c(),
  config = httr::config(),
  cas_url = NULL,
  logger = NULL
)

Arguments

url

url

serviceVersion

WFS service version

user

user

pwd

password

token

token

headers

headers

config

config

cas_url

Central Authentication Service (CAS) URL

logger

logger

Method getCapabilities()

Get WPS capabilities

Usage

WPSClient$getCapabilities()

Returns

an object of class WPSCapabilities

Method reloadCapabilities()

Reloads WPS capabilities

Usage

WPSClient$reloadCapabilities()

Method getProcesses()

Get the list of processes offered by the service capabilities. pretty allows to control the type output. If TRUE, a data.frame will be returned. When prettified output, it is also possible to get a 'full' description of the process by setting full = TRUE in which case a the WPS client will request a process description (with more information about the process) for each process listed in the capabilities.

Usage

WPSClient$getProcesses(pretty = FALSE, full = FALSE)

Arguments

pretty

pretty

full

full

Returns

a list of WPSProcessDescription or a data.frame

Method describeProcess()

Get the description of a process, given its identifier, returning an object of class WPSProcessDescription

Usage

WPSClient$describeProcess(identifier)

Arguments

identifier

process identifier

Returns

an object of class WPSProcessDescription

Method execute()

Execute a process, given its identifier

Usage

WPSClient$execute(
  identifier,
  dataInputs = list(),
  responseForm = NULL,
  storeExecuteResponse = FALSE,
  lineage = NULL,
  status = NULL,
  update = FALSE,
  updateInterval = 1
)

Arguments

identifier

process identifier

dataInputs

a named list of data inputs, objects of class WPSLiteralData, WPSComplexData or WPSBoundingBoxData

responseForm

response form, object of class WPSResponseDocument

storeExecuteResponse

store execute response? object of class logical. FALSE by default

lineage

lineage, object of class logical

status

status, object of class logical

update

update, object of class logical. For asynchronous requests

updateInterval

update interval, object of class integer. For asynchronous requests

Method clone()

The objects of this class are cloneable with this method.

Usage

WPSClient$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# \dontrun{
   #example based on a WPS endpoint responding at http://localhost:8080/geoserver/wps
   wps <- WPSClient$new("http://localhost:8080/geoserver/wps", serviceVersion = "1.0.0")
#> Error in curl::curl_fetch_memory(url, handle = handle): Failed to connect to localhost port 8080 after 0 ms: Connection refused
   
   #get capabilities
   caps <- wps$getCapabilities()
#> Error in eval(expr, envir, enclos): object 'wps' not found
# }