WPSLiteralData

WPSLiteralData

WPSLiteralData

Format

R6Class object.

Value

Object of R6Class for modelling a WPS Literal Data

Author

Emmanuel Blondel <emmanuel.blondel1@gmail.com>

Super class

ows4R::OGCAbstractObject -> WPSLiteralData

Public fields

value

value

Methods

Public methods

WPSLiteralData$new()

WPSLiteralData$decode()

WPSLiteralData$checkValidity()

WPSLiteralData$clone()

Inherited methods

ows4R::OGCAbstractObject$ERROR()

ows4R::OGCAbstractObject$INFO()

ows4R::OGCAbstractObject$WARN()

ows4R::OGCAbstractObject$encode()

ows4R::OGCAbstractObject$getClass()

ows4R::OGCAbstractObject$getClassName()

ows4R::OGCAbstractObject$getNamespaceDefinition()

ows4R::OGCAbstractObject$isFieldInheritedFrom()

ows4R::OGCAbstractObject$logger()

ows4R::OGCAbstractObject$print()

Method new()

Initializes an object of class WPSLiteralData

Usage

WPSLiteralData$new(xml = NULL, value = NULL, serviceVersion = "1.0.0")

Arguments

xml

an object of class XMLInternalNode-class to initialize from XML

value

value

serviceVersion

WPS service version

Method decode()

Decodes an object of class WPSLiteralData from XML

Usage

WPSLiteralData$decode(xml)

Arguments

xml

an object of class XMLInternalNode-class to initialize from XML

Method checkValidity()

Check the object against a parameter description inherited from a WPS process description, object of class WPSLiteralInputDescription. If not valid, the function will raise an error.

Usage

WPSLiteralData$checkValidity(parameterDescription)

Arguments

parameterDescription

object of class WPSLiteralInputDescription

Returns

an error if not valid

Method clone()

The objects of this class are cloneable with this method.

Usage

WPSLiteralData$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.