WPSProcess

WPSProcess

WPSProcess

Format

R6Class object.

Value

Object of R6Class modelling a WPS process

Note

Class used internally by ows4R

Author

Emmanuel Blondel <emmanuel.blondel1@gmail.com>

Super class

ows4R::OGCAbstractObject -> WPSProcess

Methods

Public methods

WPSProcess$new()

WPSProcess$getIdentifier()

WPSProcess$getTitle()

WPSProcess$getVersion()

WPSProcess$getDescription()

WPSProcess$execute()

WPSProcess$clone()

Inherited methods

ows4R::OGCAbstractObject$encode()

ows4R::OGCAbstractObject$ERROR()

ows4R::OGCAbstractObject$getClass()

ows4R::OGCAbstractObject$getClassName()

ows4R::OGCAbstractObject$getNamespaceDefinition()

ows4R::OGCAbstractObject$INFO()

ows4R::OGCAbstractObject$isFieldInheritedFrom()

ows4R::OGCAbstractObject$logger()

ows4R::OGCAbstractObject$WARN()

Method new()

Initializes a WPSProcess

Usage

WPSProcess$new(xml, capabilities = NULL, version, logger = NULL, ...)

Arguments

xml

object of class XMLInternalNode-class from XML

capabilities

object of class WPSCapabilities

version

service version

logger

logger

...

any additional parameter

Method getIdentifier()

Get identifier

Usage

WPSProcess$getIdentifier()

Returns

object of class character

Method getTitle()

Get title

Usage

WPSProcess$getTitle()

Returns

object of class character

Method getVersion()

Get version

Usage

WPSProcess$getVersion()

Returns

object of class character

Method getDescription()

Get description

Usage

WPSProcess$getDescription()

Returns

object of class WPSProcessDescription

Method execute()

Execute process

Usage

WPSProcess$execute(
  dataInputs = list(),
  responseForm = NULL,
  storeExecuteResponse = FALSE,
  lineage = NULL,
  status = NULL,
  update = FALSE,
  updateInterval = 1
)

Arguments

dataInputs

a named list of data inputs, objects of class WPSLiteralData, WPSComplexData or WPSBoundingBoxData

responseForm

response form, object of class WPSResponseDocument

storeExecuteResponse

store execute response? object of class logical. FALSE by default

lineage

lineage, object of class logical

status

status, object of class logical

update

update, object of class logical. For asynchronous requests

updateInterval

update interval, object of class integer. For asynchronous requests

Method clone()

The objects of this class are cloneable with this method.

Usage

WPSProcess$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.