WCSClient

WCSClient

WCSClient

Format

R6Class object.

Value

Object of R6Class with methods for interfacing an OGC Web Coverage Service.

Author

Emmanuel Blondel <emmanuel.blondel1@gmail.com>

Super classes

ows4R::OGCAbstractObject -> ows4R::OWSClient -> WCSClient

Methods

Public methods

WCSClient$new()

WCSClient$getCapabilities()

WCSClient$reloadCapabilities()

WCSClient$describeCoverage()

WCSClient$getCoverage()

WCSClient$clone()

Inherited methods

ows4R::OGCAbstractObject$ERROR()

ows4R::OGCAbstractObject$INFO()

ows4R::OGCAbstractObject$WARN()

ows4R::OGCAbstractObject$encode()

ows4R::OGCAbstractObject$getClass()

ows4R::OGCAbstractObject$getClassName()

ows4R::OGCAbstractObject$getNamespaceDefinition()

ows4R::OGCAbstractObject$isFieldInheritedFrom()

ows4R::OGCAbstractObject$logger()

ows4R::OWSClient$getCASUrl()

ows4R::OWSClient$getConfig()

ows4R::OWSClient$getHeaders()

ows4R::OWSClient$getPwd()

ows4R::OWSClient$getToken()

ows4R::OWSClient$getUrl()

ows4R::OWSClient$getUser()

ows4R::OWSClient$getVersion()

Method new()

This method is used to instantiate a WCSClient with the url of the OGC service. Authentication is supported using basic auth (using user/pwd arguments), bearer token (using token argument), or custom (using headers argument). By default, the logger argument will be set to NULL (no logger). This argument accepts two possible values: INFO: to print only ows4R logs, DEBUG: to print more verbose logs

Usage

WCSClient$new(
  url,
  serviceVersion = NULL,
  user = NULL,
  pwd = NULL,
  token = NULL,
  headers = c(),
  config = httr::config(),
  cas_url = NULL,
  logger = NULL
)

Arguments

url

url

serviceVersion

WFS service version

user

user

pwd

password

token

token

headers

headers

config

config

cas_url

Central Authentication Service (CAS) URL

logger

logger

Method getCapabilities()

Get WCS capabilities

Usage

WCSClient$getCapabilities()

Returns

an object of class WCSCapabilities

Method reloadCapabilities()

Reloads WCS capabilities

Usage

WCSClient$reloadCapabilities()

Method describeCoverage()

Describes coverage

Usage

WCSClient$describeCoverage(identifier)

Arguments

identifier

identifier

Returns

an object of class WCSCoverageDescription

Method getCoverage()

Get coverage

Usage

WCSClient$getCoverage(
  identifier,
  bbox = NULL,
  crs = NULL,
  time = NULL,
  format = NULL,
  rangesubset = NULL,
  gridbaseCRS = NULL,
  gridtype = NULL,
  gridCS = NULL,
  gridorigin = NULL,
  gridoffsets = NULL,
  method = "GET",
  filename = NULL,
  ...
)

Arguments

identifier

Coverage identifier. Object of class character

bbox

bbox. Object of class matrix. Default is NULL. eg. OWSUtils$toBBOX(-180,180,-90,90)

crs

crs. Object of class character giving the CRS identifier (EPSG prefixed code, or URI/URN). Default is NULL.

time

time. Object of class character representing time instant/period. Default is NULL

format

format. Object of class character Default will be GeoTIFF, coded differently depending on the WCS version.

rangesubset

rangesubset. Default is NULL

gridbaseCRS

grid base CRS. Default is NULL

gridtype

grid type. Default is NULL

gridCS

grid CS. Default is NULL

gridorigin

grid origin. Default is NULL

gridoffsets

grid offsets. Default is NULL

method

method to get coverage, either 'GET' or 'POST' (experimental - under development). Object of class character.

filename

filename. Object of class character. Optional filename to download the coverage

...

any other argument to WCSGetCoverage

elevation

elevation. Object of class character or numeric. Default is NULL

Returns

an object of class SpatRaster from terra

Method clone()

The objects of this class are cloneable with this method.

Usage

WCSClient$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# \dontrun{
   wcs <- WCSClient$new("http://localhost:8080/geoserver/wcs", serviceVersion = "2.0.1")
#> Error in curl::curl_fetch_memory(url, handle = handle): Failed to connect to localhost port 8080 after 0 ms: Connection refused
# }